Rename EditingState to SavedEditingState, plus other minor improvements.
Rename EditingState to SavedEditingState to better reflect the fact that
it is a collection of members I happen to choose to save for hidden
buffers, but are not otherwise remarkable.  In particular, it is nowhere
near *all* of the editing state.

Add comments explaining the differences among Buffer, HistoryBuffer,
SavedEditingState, BufferState, and Editor.

SavedEditingState: Add comments better explaining the role of its
members.

editor.h: Explain the embedded SavedEditingState.

editor.cc: Add many explicit "this->" qualifiers, although nowhere near
all possible.  I may or may not complete the transformation, but I think
the changes so far enhance readability.  Also marked a number of
variables in Editor::updateFrame as 'const', again for improved
readability.