main.cc: Fix scroll bug: thumb would misbehave near end of file.
The symptom was I would page down to the end of the file, and as soon as
I went past the end, the thumb would jump up to one page before the end,
even though everything else indicated I was at the end of the range.

The bug was caused by setting the new value before the new range,
causing the value to be clamped to the old range.