main.cc: Place a thin bkack border around the main editor widget.
The Qt3 code put a border around the menu bar, but that is not
encouraged with Qt5 because QMenuBar no longer inherits QFrame.  So I
chose to put a border around the editor so it is thinly divided from the
menu as well as the scrollbar and status bar.

Of course, the way this looks is heavily influenced by both my choice of
colors in the editor and the extremely spartan design of the Windows 10
menu and scrollbar widgets.  Anyway, it is adequate as a substitute for
the menu border I had with Qt3.