Fix recognition of Window menu entries, using ID instead of title.
This should make things work properly even when there are multiple
files open with the same basename, although I have not tested that.

This is more like how it worked with Qt3.  I didn't see how to do it in
Qt5 at first, but now I see that the QAction::data() field is a
QVariant, which can hold a numeric ID.

I also removed the awkward 'recentMenuBuffer' member.  It did not work,
since Qt5 evidently fires the events in the opposite order.  But it is
not needed, and I don't see why it was originally either.