Fix search "not found" box appearing in the wrong place.
The bug was due to specifying Qt::ToolTip, which evidently causes the
window position to be interpreted as relative to the screen rather than
the parent window (or perhaps the parent is ignored entirely?).  Anyway,
it seems I should not use Qt::ToolTip myself (which the documentation
hints at).

I also had to arrange for an opaque background, and chose to add a
little more horizontal space around the text inside the label.