Fix bug: Weird vertical bars when searching in file with no newline.
The fix is to ensure, while painting, that we do not check for search
hits in lines beyond the end of the file.  When we do, 'findString'
clamps the starting line/col to be valid, and then it can find spurious
hits because it is searching in the wrong line.  Normally that is not a
problem, but if the file has no newline, then the last "line" as the
editor core sees it is not empty, and hence can contain hits after
clamping the position.  (When the last line is empty, the search always
fails to find any hits.)

Contributing to the bug is the questionable behavior of 'findString'
clamping its input position.  That seems to be mainly due to
'walkCursor' refusing to work with out of bounds positions.  One of
those two probably should deal better with this case, but for now it is
easy enough to avoid it altogether.