Expand line/col status box a little and change default window position.
Making the status box 80 pixels wide (instead of 60) ensures that the
'*' indicating unsaved changes is visible even with (at least) 6 digits
of line number and 3 digits of column number displayed.

As it happens, this change makes the status bar slightly wider than the
previous default width of 400 when the incremental search is active, so
I went ahead and moved the window to (400,100) with size 600x800 by
default.  Obviously, long term, I need to make that configurable and
persistent, but it is slightly better than it was.