Fix bug: Margin was painted with selection background if multiple lines selected.
Effectively, the selection style was being retained as we wrapped around
to the next line.  There is a certain logic to that, but especially with
the one-pixel margins I currently have set, it looked like a bug.  So
now the left margin is always painted with the normal background color.