Fix bug: Bytes in [128,255] were displayed as blank.
This is due to being careless about character encoding from the start. I
added some comments explaining that but haven't fixed anything, other
than forcing the interpretation to be Latin-1 when painting the window.

history.cc: I fixed what appears to be a latent non-portability in
se8bit due to assuming that 'char' is signed.