main.cc: fileOpen: Add comment with way to disable native dialog.
I have been having trouble with the native Windows file-open dialog
being slow to open and causing my desktop icons to refresh.  I tried
disabling that, but even with the non-native dialog, the same thing
happens!  So I reverted the change but I'm leaving it commented so I
can play with it more later maybe.