Reimplemented undo/redo to ignore cursor movement.
This is preliminary to being able to get rid of CursorBuffer since
I need to change where in the class hierarchy the cursor is stored,
but that will mean that undo/redo will not see it.

It also makes undo/redo work more like I want it to, since undoing
cursor movement was kind of useless and annoying.