Add TextDocumentEditor class, approximately replacing Buffer.
This is the main piece of the major refactoring I am doing, placing
a class between EditorWidget and TextDocument.  It turns out most of
the code that is now in TextDocumentEditor was originally in Buffer
(which has now been removed), but the data organization is new.

See doc/classes.ded for the "before" picture and doc/classes-new.ded
for the "after" (as they exist in this commit; I plan to rename them).

Detailed changes:

Delete buffer.{h,cc}, create text-document-editor.{cc,h}.

Move test code from text-document.cc to new test-text-document-editor.cc
program.

Remove SavedEditingState.  The selection (mark) and first/lastVisible
(scrolling) fields went into TextDocumentEditor while the hitText fields
went into EditorWidget.

I added selfCheck routines to EditorWidget and TextDocumentEditor.

A number of moved fields were replaced with methods of the same name.
Most of this is meant as temporary scaffolding as I continue this
refactoring.

Rename HBGrouper to UndoHistoryGrouper.

EditorWidget now has a set of TextDocumentFileEditor objects to keep
track of per-file and per-widget data.  It consists primarily of the
TextDocumentEditor itself, each of which is now owned by a specific
editor widget.  This replaces the old system where a widget would store
its cursor and scroll state in a SavedEditingState associated with the
file (and shared by all widgets).  The collection of TDFE objects is
stored in EditorWidget::m_editors.