test-td-editor.cc: Kill mutants 547 and 560 in findString.
Mutant 560 is interesting because killing it requires corrupting memory.
I spent a little time looking into modifying GrowArray to maintain a
canary, but the trouble is it has to work with any type T, so I would
need both a canary value and a way to check for it for an arbitrary
type, which is challenging.  Or I could drop down to a lower level of
abstraction, but that has issues too.