EditorWidget: rename all data members to start with "m_".
Also replace all resulting "this->m_" with just "m_".

I'm not completely sure I will ultimately stay with this convention, but
the limited places I have used it seem to be improved by it, so I want
to try it out in a wider setting than I have so far.