Fix bug: Unsaved-changes marker lagging in other windows.
The basic bug was that no notification was sent explicitly for a state
transition of unsaved-changes true to false or vice-versa.  In most
cases, other notifications would happen to cause the display to update,
but one keystroke too late.  In a few (e.g, undo) the sequencing
happened to cause the update to arrive in sync, but that was fragile.

The solution is to add another notification event, slightly abusing the
existing TDO interface to piggy-back TD information on an interface that
otherwise only deals with TDC.