Start implementing Ctrl+I.  But also start using RCSerf.
What happened here is I wrote part of Ctrl+I, and immediately caused a
new memory corruption error.  I've been doing that more than I would
like to, so I went on a detour and built RCSerf (in smbase).  Using that
back in editor, it was able to detect and prevent the corruption.  So
now I want to check in what I have and go through and use RCSerf
everywhere it makes sense.