Refactor/assertion: Assert name is set in NTDL::addDocument.
I decided not to require it in the NTD constructor because there are
two options (file or non-file) and relying purely on overloading loses
the self-documenting benefit of 'setFilename' and 'setNonFileName'.
Instead, I assert that the name has been set in addDocument() so I can
be sure one of the 'set' functions was called.