Create new TextSearch module.  Use it to show number of offscreen search matches.
This is a first step toward being able to do more advanced forms of
searching and feedback.  It incrementally maintains search hits and
gets calls to the search algorithm out of the paint(!) code.

I also still need to rip out and throw away the old 'findString'
implementation in TextDocumentEditor.

ogap.h: This module is now used.