Rename TDC::getTextSpan to getTextSpanningLines.  Same for TD (unused).
I also decided against removing this method, so removed the TODO.  It is
a little weird to have this and also getTextRange, but it is called from
two places for which changing to getTextRange would not really be an
improvement.  So, with a longer name, I now think it's fine.