editor-window: Add reloadCurrentDocument using VFS.
editor-widget: Call window's reload instead of doing it directly.

In EditorWindow::reloadCurrentDocument, the need to set
m_ignoreTextDocumentNotifications is a bit of a mystery.  Without it,
the widget scrolls the view to the end upon reload.  What I do not
understand is why this wasn't needed before, when using direct file
system access, since all the code paths look the same.