named-td-list: Use 'DocumentName' instead of 'string' as the identifier.
This part of an effort to add a host name to DocumentName and then use
DocumentName like I had used the string before.

doc-name: Rename m_docName to m_resourceName.

This goes with the above because it's confusing to have two kinds of
"document name".

named-td: Add documentName and setDocumentName in support of the above
changes.

Other modules: Push the change to DocumentName through the interfaces
that name documents, as well as the rename to 'resourceName'.