editor-widget.cc: Fix bug: Assertion failure while resizing window.
The assertion failure was due to accessing 'text' at an index equal to
its length.  The added condition prevents that, and is symmetric with
the check for negative.

Something else must also be doing a similar check since the crash was
difficult to reproduce, but I didn't find that check.