editor-widget: Use new `DiagnosticDetailsDialog` to show diagnostics.
This is very much work in progress, but this crude first cut has most
of the required functionality, so is worth checkpointing.

Add `diagnostic-details-dialog` module.  This version closely follows
what ChatGPT provided when given the spec as its prompt, although I
retyped almost all of it and made a few adjustments along the way.