Move `NamedTextDocument::m_lspWaitingForDiagnosticsForVersion` to `LSPDocumentInfo::m_waitingForDiagnostics`.
This is part of an effort to move LSP knowledge out of `NTD` and
keep it in the LSP manager.

Remove `lsp-doc-state` module.  It mixes several concerns, which is
what the LSP status bar does too, but no single component is really
in a position to deliver this data.

lsp-status-widget: Re-do how status is computed, relying more heavily
(but not exclusively) on what the LSP manager knows.