textmcoord-map, td-diagnostics: Add `m_numLines`.
This is a large change motivated by `test_deleteNearEnd` in
`textmcoord-map-test`.  The issue is the diagnostics observer
promises to ensure all diagnostics coordinates are valid, but
cannot do that when the last line is deleted without being
able to recognize that.

There are some loose ends and questionable elements, but all
the tests pass so this seems like a good place to commit before
making further changes.