editor-global: Fix: ApplyCommandDialogs break test retry, etc.
In editor-window, I had been keeping two `static` pointers to the
`ApplyCommandDialog`s.  But this breaks the `-testCommands` option
since that creates a new `EditorGlobal` for each test.  One
consequence is that retry gets into a weird state where Qt does
not work.  Another is that running two tests that both open one
of the dialogs fails.

The lesson here is to not be lazy with `static` locals, and instead
properly put everything long-lived into `EditorGlobal`.