diagnostic-details-dialog: Add a window icon.
This is just a screenshot of an actual dialog with approximately
square aspect ratio, scaled down to 64x64 with no interpolation.
It's only visible in the upper-left corner of the dialog, since
these don't appear in the taskbar, so of little relevance, but I
saw it previously had the default icon and decided to change it.

Other dialogs still mostly have the default, which is fine.