editor-widget: mouse{Press,Move,Release}Event: Do not call base class method.
The base class method calls `ignore()` on the event object, which
causes it to propagate up the tree.  Those events then interfere
with the recording.

I had a comment "get rid of popups?" but I don't know what that
meant, and removing the call does not cause any evident problems.