event-replay: Parse the .ev file as one GDValue rather than many.
slot_aboutToBlock: Remove the EOF check.  For most tests it is simply
irrelevant, but for empty.ev, it actually prevented the test from
completing since we only break the loop if we post the quiescence
event.  Previously that worked because EOF was not detected right away
when using the istream.  So it was unnecessary, wrong, and accidentally
masked before.

smbase: GDValues now have source location information.  That is what
enables this change in how event replay works.

editor-global: innerMain: Push the file name on the exception context
stack so it gets reported.  That previously happened due to how the
reader reports errors.

test/*.ev: Surround all commands with brackets to make a proper GDVN
sequence.